refactor(palette): split shared type modules
